Autod Procejt 3.3 up to date
1. find the reason did allow us automatically to insert id from table Landing to Carservice table in column customerid.

1.1 action with the help which we have achieved to desired functionality.

1.
creating a CarServiceDto without specifying a CustomerId, and then retrieving the corresponding LandingPage and installing the CustomerId in CarServiceDto

2.
Relationship between LandingPage and CarService: A modification to reflect the processing of relationships between LandingPage and CarService. The SaveCustomerRequest action installs the CustomerId in the CarService based on the corresponding LandingPage before calling the SaveCustomerRequest method.

3.
Database Lookup for LandingPage: By using_autoContext.LandingPages.FindAsync(vm.CustomerId), we perform a database lookup to ensure that the LandingPage with the specified CustomerId exists before proceeding.

2. Now after we fill the first primary data form we redirected to the form where a user input his car and service he needs, column customerId, get id from the first form letting us to know what customer has filled both form.